replies-pipeline/requirements.txt
elasticsearch==7.10.0
kafka-python==2.0.2
matplotlib==3.3.3
nltk==3.5
numpy==1.19.4
pandas==1.1.4
pickleshare==0.7.5
pqdict==1.1.1
psycopg2-binary==2.8.6
scikit-learn==0.23.2
scipy==1.5.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
transformers==4.1.1
# vaderSentiment==3.3.2
